Overhaul furnace
This submission replaces the furnace with a copy of the default furnace
that's been modified to work with the crafting mod's crafting system.

In the process I've fixed some omissions in the legacy.lua recipe
importer and added some more "atomic" operations to util.lua.

The mod also needed a default dependency added. This was always implicit
in its use of default textures, but is now explicit due to formspec
requirements.